Detect timeout of serial transmission
The SerialPort will now throw an exception when the serial line cannot
be read (or written) for more than two seconds.

This limit does not apply to the underlying QSerialPort (which is used
to read and write the serial console), but only to the bullycpp backend
code which uses the ISerialPort interface.

As far as I can tell, PicBootloaderDriver::programHexFile() should be
strongly exception safe.

Fixes #1.