Add simple console wrapper
This program is built only on Windows and simply forwards output from
the child process.  On Windows, the main executable is renamed
BullyCPPw; the new wrapper invokes it.  Users should not notice a
difference if they have BullyCPP.exe shortcuts or automation, other than
proper output on the command line.

Closes #16 on GitHub.